+++ packages/eslint-markdown/src/configs/all.js
@@ -49,7 +49,6 @@ export default function all(plugin) {
       'md/consistent-strong-style': 'error',
       'md/consistent-thematic-break-style': 'error',
       'md/en-capitalization': 'error',
-      'md/heading-id': 'error',
       'md/no-bold-paragraph': 'error',
       'md/no-control-character': 'error',
       'md/no-curly-quote': 'error',
@@ -59,6 +58,7 @@ export default function all(plugin) {
       'md/no-irregular-dash': 'error',
       'md/no-irregular-whitespace': 'error',
       'md/no-url-trailing-slash': 'error',
+      'md/require-heading-id': 'error',
       'md/require-image-title': 'error',
       'md/require-link-title': 'error',
     },
