+++ packages/eslint-markdown/src/rules/index.js
@@ -10,7 +10,6 @@ import consistentEmphasisStyle from './consistent-emphasis-style.js';
 import consistentStrongStyle from './consistent-strong-style.js';
 import consistentThematicBreakStyle from './consistent-thematic-break-style.js';
 import enCapitalization from './en-capitalization.js';
-import headingId from './heading-id.js';
 import noBoldParagraph from './no-bold-paragraph.js';
 import noControlCharacter from './no-control-character.js';
 import noCurlyQuote from './no-curly-quote.js';
@@ -20,6 +19,7 @@ import noGitConflictMarker from './no-git-conflict-marker.js';
 import noIrregularDash from './no-irregular-dash.js';
 import noIrregularWhitespace from './no-irregular-whitespace.js';
 import noUrlTrailingSlash from './no-url-trailing-slash.js';
+import requireHeadingId from './require-heading-id.js';
 import requireImageTitle from './require-image-title.js';
 import requireLinkTitle from './require-link-title.js';
 
@@ -34,7 +34,6 @@ export default {
   'consistent-strong-style': consistentStrongStyle,
   'consistent-thematic-break-style': consistentThematicBreakStyle,
   'en-capitalization': enCapitalization,
-  'heading-id': headingId,
   'no-bold-paragraph': noBoldParagraph,
   'no-control-character': noControlCharacter,
   'no-curly-quote': noCurlyQuote,
@@ -44,6 +43,7 @@ export default {
   'no-irregular-dash': noIrregularDash,
   'no-irregular-whitespace': noIrregularWhitespace,
   'no-url-trailing-slash': noUrlTrailingSlash,
+  'require-heading-id': requireHeadingId,
   'require-image-title': requireImageTitle,
   'require-link-title': requireLinkTitle,
 };
